Add files via upload
The shapes are for now just hard coded onto the window in the constructor, but once we connect the backend, we can switch that around easily.  The buttons have empty listeners, with the exception of the quit button since that was an easy listener to do.  The methods and comments should be good, with the exception of the listeners of the other 9 buttons.  I also included the Input class that has the main.